move controls things from index.ts
